refactor: json to `deployed.ts`
