+++ src/template/deployed-template.ts
@@ -1,0 +1,11 @@
+export function getDeployedSource(deployedJson) {
+  return `
+interface DeployedAddress {
+  [chainName: string]: {
+    [contractName: string]: string
+    }    
+}
+
+export const deployed: DeployedAddress = ${JSON.stringify(deployedJson, null, 2)} 
+`
+}
